% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportStackObj.R
\name{exportStackObj}
\alias{exportStackObj}
\title{Export the Object Stack}
\usage{
exportStackObj(stack, tag)
}
\arguments{
\item{stack}{The object stack where the exporting images of individual object is from}

\item{tag}{The name of the experiment or anything that identifies the experiment or slide. It should a string that does not contain space in between}
}
\value{
The file exports the images of individual object identified that are stored in the image stack as tiff file with bits.per.sample set to 16
}
\description{
massive export of image of objects
}
