% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_object_stack.R
\name{gen_object_stack}
\alias{gen_object_stack}
\title{Generating Object Stack}
\usage{
gen_object_stack(label_image, image)
}
\arguments{
\item{label_image}{The object identified labeled image stack}

\item{image}{The original image that the labeled image was generating from}
}
\value{
A Image class image stack that contains the obejct identified
}
\description{
generating object stack. This code wraps the EBImage function stackObjects in order to make the function compatible with the massive images processing
}
