% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3_channel.R
\name{get_3_channel}
\alias{get_3_channel}
\title{Get 3 channel}
\usage{
get_3_channel(image_vector)
}
\arguments{
\item{image_vector}{The original color image stack}
}
\value{
The single 3rd channel only image
}
\description{
create the frame 3 for all the images in the image stack that was imported
}
