% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_images.R
\name{import_images}
\alias{import_images}
\title{Import Images}
\usage{
import_images(directory)
}
\arguments{
\item{directory}{The directory that contains the images that user will be working on}
}
\value{
The function returns and stores a Image class object in the Global Environment that contains the images user choose to import
}
\description{
import all the images from the working directory into a image stack. this code wraps the orginal readImage from EBImage package (4.22.1) with user-friendly interactions to choose the files and make massive processing of images feasible
}
