% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MgMorphoClass}
\alias{MgMorphoClass}
\title{Main user interface that runs the whole pipeline}
\usage{
MgMorphoClass()
}
\description{
This is to make the functions that present in the package be easier to use.
}
