% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_image.R
\name{processing_image}
\alias{processing_image}
\title{Processing the image stack}
\usage{
processing_image(image_vector)
}
\arguments{
\item{image_vector}{The single channel image stack need to be processed}
}
\value{
The results from each step of the processing
}
\description{
This function is to process the image stack, identifies the object from the images and measures the characteristic of the objects.
First thing that fucntion do is adjust all images to be at the same brightness. Using the first image as the reference and adjust the median of intensity of the image to be same as median of intensity of the first image
Afterward, the function convert the image into a negative image and filter out the pixel with intensity less than 0.72. The function then fill the hull in the object.
The image then subjected for object identification and paint the original image with circling around the object. The identified objects will be measures then.
}
