% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_selection.R
\name{random_selection}
\alias{random_selection}
\title{Random selection of image object from image stack}
\usage{
random_selection(image_stack, seed)
}
\arguments{
\item{image_stack}{the object stacks that the user is going to work on. The import should be a string containing name of object(s) in GlobalEnvironment}

\item{seed}{The seed to be set for the randomnization. This is an optional parameter intending to create the reproducibility of the randomnizaiton process}
}
\value{
selected frames images in an image stack and the number of the frame in original stack from each of the stacks user input
}
\description{
To establish the training set for cell type classifier, a pre-determined classification of object has to be perfermed.
In order to get a random set of image for manual classification which will be used as training set for later machine learning steps.
}
