% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_cell_type.R
\name{record_cell_type}
\alias{record_cell_type}
\title{Record cell type}
\usage{
record_cell_type(image_stack)
}
\arguments{
\item{image_stack}{The image stack that the user is going to determine the class of microglia activation status}
}
\value{
A data matrix recorded the type of microglia activation status that the user decided
}
\description{
This function/ program is for user to perform the classification of objects that is randomly selected.
}
